Rename setup_java.py to java.py
