+++ tests/test_scripting.py
@@ -1,6 +1,6 @@
 from magicgui import magicgui
 
-from napari_imagej.setup_imagej import JavaClasses
+from napari_imagej.java import JavaClasses
 from napari_imagej.utilities._module_utils import functionify_module_execution
