+++ tests/types/test_converters.py
@@ -6,7 +6,7 @@ from jpype import JArray, JDouble
 from labeling.Labeling import Labeling
 from napari.layers import Labels, Points, Shapes, Surface
 
-from napari_imagej.setup_imagej import jc
+from napari_imagej.java import jc
 from napari_imagej.types.converters.labels import _labeling_to_layer, _layer_to_labeling
 from napari_imagej.types.standins import OutOfBoundsFactory, StructuringElement
 from napari_imagej.types.type_conversions import python_type_of
