+++ tests/widgets/test_focuser.py
@@ -1,7 +1,7 @@
 import pytest
 from qtpy.QtWidgets import QLabel, QVBoxLayout, QWidget
 
-from napari_imagej.setup_imagej import JavaClasses
+from napari_imagej.java import JavaClasses
 from napari_imagej.widgets.focuser import FocusWidget
 from napari_imagej.widgets.layouts import FlowLayout
